Experimental code -- Learning how Promises and chaining works. ----------------------------------------------------------------- modified -- HomeView.vue The previous code was moved to DevView.vue.
This file contains a simple UI and a bunch of TypeScript.

This was taken from Mdn web site and tweaked to work in Vue3 - Composition, TypeScript and Vuetify.

NOTE -- This file contains a TON of debugging comments and console.log() statements. -- They are essential to understanding how the 'flow' works.