Solved Validation + Submit using Promise() technique ---------------------------------------------------- This solution has a SINGLE <v-text-field> that is validated with 3x Promise style validators. This technique uses a "Promise.all( [ ] )" function for handling multiple validation promises.
The results of the Promis.all() is passed to the final ".then()" as a <string[]> object.
The final ".then()" handles processing the results <string>[] and returns a single <boolean> true or a single validation-error-message string.

A key point is that VALIDATION and the displaying of error messages, is handled in VALIDATION not SUBMIT.
The overall validation success/failure results is tested as a boolean in the @submit function (myFunctionSubmit()).
There is NO validation data available in the SUBMIT function.