Add new Phone Number form field with 2x validators. --------------------------------------------------- modified -- HomeView.vue
Add checkPhoneMissingInvalidChar() validator.
---------------------------------------------
Tests phone number to see if it is empty.
Tests the phone number to see it contains alphabetical characters.

Add checkPhoneFormatAreaCode() validator.
-----------------------------------------
Tests the phone number for a minimum of area code, exchange and line number. EG: "919 272 7866".
-- Has provisions for hyphen and space.
-- Has provision for parenthesis around area-code.
-- Has provision for optional country-code '1' with area-code, exchange and line-number.
Tests the phone number for a '919' area-code.

NOTE
----
This code contains a lot of comments and debug code that is designed to help understand what happens during validation.

The code also contains many TypeScript 'warnings'. These warnings don't cause the app to fail but should, eventually, be cleared up.