version -- v3-auth-form-update-v1 ----------------------------------------------------------- Add a .Then() to the Auth.currentUserInfo() Promise.  * Use the .Then() to update the nicknameModel.
Add Auth.currentAuthenticatedUser().then() to the hub.listener().
 * The .then() will be used to update the emailModel.
 * This is required because the listener payload does not contain the email address. - It must be fetched sepparatly.

Modified/Renamed getNickname() to getNicknameEmail()
 * The function now returns two objects: nicknameModel and emailModel.
Add call to getNicknameEmail() whenever the page refreshes or is updated.
---
Add nicknameModel and emailModel to the new info section.
Relayout the {{ user.attribute.* }} info.
Relayout the debug emails
Add a bunch of debugging code to track the sync and async functions.
---
The most of the debugging code and commented out lines will be removed in the next commit.