version -- v3-auth-form-update-v3 -----------------------------------------------------------
--- BUG FIX
In the SignIn work flow, the nicknameModel and the Cognito nickname would be deleted.
What was happening was that an empty nicknameModel was being updated in Cognito on each SignIn.
Later, still in the SignIn workflow, auth would not find and return the nickname because it had be overwritten.
The nicknameModel only exists during SignIn when it is part of the SignUp workflow.

--- SOLUTION
The solution tests to see if the nicknameModel exists when Signing In.
The nicknameModel only exist during SignUp workflow.
Checking to see if the nicknameModel exists solves the problem.
IF -- nicknameModel exist, THEN -- Add it to Cognito.
The next step will return the nickname along with the other user attributes.

-----------------------------------------------------------
Modified -- HomeView.vue

-----------------------------------------------------------
Description

Replace Hub.listen > case "signIn" code
 * Check to see if nicknameModel exists.
 * -- Add to Auth user if true.
 * -- This only happens during SignUp.

Moved unused Hub.listen > events to top and comment out.
 * Keep these for future reference.

Replace 'chained' assignment statements with separate statements.
 * There is something about this that doesn't seem to be working.
 * This deserves more research.

Delete blank lines that are appropriate.
Delete debug and commented out code.