version -- v3-auth-pool-v21 -------------------------------------------------- Add code to trap error when adding a preferred_username fails. This failure happens when the username is already used.
Comments, debug and obsolete code, will be discarded in the NEXT commit.

--------------------------------------------------
modified:    src/views/  ProfileView.vue
--------------------------------------------------

Add v-dialog that opens when the preferred_username fails to update because the username is already being used.

Add const invalidUsernameDialogFlag for controlling the display of the dialog.

Add debug code to prevent the old validation from checking the requested preferred_username.
____ This change and all of the code that does the "AdminGetUserCommand()" in validation will be discarded in the next commit.

--------------------------------------------------