version -- v3-auth-vtfy3-v5 --------------------------------------------------
This file contains a working example of v-dialog implemented inside the
Account Confirmation Ui.

See the "Resend Code Dialog" block, which is commented out.

Uncomment the code and the dialog should work.

This file is for reference ONLY.

It will not be included in the current project.

--------------------------------------------------
added:    src/views/  HomeView_Dialog.vue
--------------------------------------------------