version -- v3-auth-vtfy3-v12 --------------------------------------------------
Work-In-Progress -- WIP

This collection of changes all relate to redesigning the HomeView.

A v-card was added beneath each component.
The card controls the show/hide of each of the components.
The card displays the Model it is related to.
When the component is opened, the related model is displayed for reference.
When the associated model has been updated, the component closes and the card is redisplayed with the updated Model.

Emits, listeners, handlers and props are used to pass important data between the child components and the HomeView.

This change allowed the 'User Info' <template> block to be removed.
The overall page is more compact.
Responsive behavior is better.

There is still more work to be done.
Each child component needs:
1) The <template> block needs cleanup.
2) The v-cards in HomeView should be moved to each child component.

--------------------------------------------------
modified:    src/views/        HomeView.vue
modified:    src/components/   Email.vue
modified:    src/components/   Nickname.vue
modified:    src/components/   Password.vue
modified:    src/components/   Phone_number.vue
modified:    src/components/   Preferred_username.vue
--------------------------------------------------

--------------------------------------------------