version -- v3-auth-vtfy3-v12 --------------------------------------------------
Work-In-Progress -- WIP

This is a new component.
This component will eventually replace DeleteUser.vue component.

DeleteUserConf.vue is a variation of DeleteUser.vue.
It has additional code that requires confirmation BEFORE the user may be deleted.
It contains a Confirmation Ui with Resend code.
It contains an Invalid confirmation code input Ui.
The confirmation APIs have been tested and function.
   They are currently disabled in this commit to prevent AWS throttling.

NOTE
----
The Confirmation Ui message needs to be added.
This will be in the next commit.

--------------------------------------------------
added:    src/components/   DeleteUserConf.vue
--------------------------------------------------