version -- v3-auth-vtfy3-v14 --------------------------------------------------
Work-In-Progress -- WIP

This is the First version of the 'burger-menu' that has potential.
This uses
   v-navigator-drawer
   v-list
   v-list-group
   <template v-slot:activator={" props "}>
   <v-list-item v-bind="props">

There are debug menu/sub-menu titles and icos in place.
These will be replaced.

NOTE
----
There is a lot of commented-out code that that will remian for reference,
It will be removed later.

--------------------------------------------------
modified:   src/layouts/        MasterLayout.vue
--------------------------------------------------