version -- v3-auth-vtfy3-v14 --------------------------------------------------
Work-In-Progress -- WIP

This is commit 2 of 2.
It is related to the rebuild of the Menu system.

There are simply too many changes to describe.
The 'old' <template> block has been almost completly replaced.

New const decls have been added to support submenus.

In general, this new design should be easier to read and extend.

NOTE
----
There are some Burger-Menu behavior issues that have not been solved.
The MOB and DSK view breakpoints have not been added yet.

--------------------------------------------------
modified:   src/layouts/   MasterLayout.vue
--------------------------------------------------