version -- v3-auth-vtfy3-v14 --------------------------------------------------
Add/Restore menu breakpoints.
Based on the display width, menues will be visible or not visible.
Made a number of layout changes to make the best use of the MOB and DSK space.

>>>--->   MasterLayout.vue

Changed the v-switch to a v-icon for toggling the theme.
   When the Cx changes the theme the icon changes.
   ___ 'mdi-weather-sunny' or 'mdi-weather-night'.
Moved the v-icon to the right side of the menu bar.
Abreviated the App Title display from 'v3-auth=vtfy3-v14' to '( v14 )'
   The DSK view will display the full title.
   The MOB view will display the version only of the title.

>>>--->   Vuetify.ts

Add "display: { mobileBreakpoint: 'sm' }"
   This was added so I could modify the 'mobile' breakpoint.
   The value is set, by default, to 'md', which is too wide for my app.

--------------------------------------------------
modified:   src/layouts/      MasterLayout.vue
modified:   src/plugins/      vuetify.ts
--------------------------------------------------