Version:   v3a-theme-pinia-v2 Date:      Thursday, August 31, 2023 @ 12:17pm --------------------------------------------------
The App-Bar icon:Theme does not change based on the active theme.
Remove the icon:Theme feature that dims and makes the icon unavailable when signed out.

--------------------------------------------------
modified:   src/layouts/      MasterLayout.vue
--------------------------------------------------

Modify the <template> Theme switch.
   Greatly simplified.
   Removed the dimming.
   Removed the conditional check that enables the @click
NOTE
----
This is interesting code.
It uses v-on / event listener to control the presence or absense of the click event.
   v-on="piniaStore.connected ? {click: onChangeSwitch} : {}"

Modify onChangeSwitch(~)
   Add code that updates piniaStore.activeTheme based on the theme.global.name.value.

Replace const themeIcon with computed value based on the piniaStore.activeTheme.
   The themeIcon changes toggles between 'mdi-weather-sunny' and 'mid-weather-night'

--------------------------------------------------