Version:   v3a-theme-pinia-v2 Date:      Tuesday, September 13, 2023 @ 4:19 pm --------------------------------------------------
Fix issue with the icons and colors not changing as expected, in the ThemeChanger.

--------------------------------------------------
modified:   src/components/   ThemeChanger.vue
--------------------------------------------------

modify toggleTheme(~)
   Discard most of the existing code and rewrite.
   Define different code based on which icon:btn:{theme} is selected.
   ___ There are two 'groups': Icon & color and Global Theme.
   ___ Code related to Active and Inactive Theme was discarded.
   ___ Code related to 'flipping' A_Theme and B_Theme was discarded.

NOTE
----
When a Theme is GREEN with switch == on, and the Theme is selected, no changes are made.
When a Theme is RED with switch == off, and the Theme is selected, the global theme changes -- The colors and icons will 'toggle'.