Version:   v3a-theme-pinia-v2 Date:      Thursday, September 14, 2023 @ 9:07 am --------------------------------------------------
Fix issue with the MasterLayout and the ThemeChanger being out of sync.

DETAIL
------
Toggling the Theme using the MasterLayout > App-Bar was not synced with the ThemeChanger.

--------------------------------------------------
modified:   src/layouts/      MasterLayout.vue
modified:   src/components/   ThemeChanger.vue
--------------------------------------------------

   >>>--->   ThemeChanger.vue

Add code to initialize the ThemeChanger when the page, containing this component, is loaded/Reloaded.

   >>>--->   MasterLayout.vue

Modify onChangeSwitch(~)
   Rename -- toggleTheme(~)
   Add code to set ALL of the Pinia Store vars used by ThemeChanger
   ___ This happens when the App-Bar > icon:weather is selected.

--------------------------------------------------