Version:   v3a-todo-theme-v1 Date:      Wednesday, October 26, 2023 @ 3:33 pm --------------------------------------------------
Fix validation and clearable features in form > input.
   The 'clearable' feature was generating an ChromeDev exception.
   The input must be validated in this order:

      value => !value ? 'Required' : true,
      value => !!value,
      value => value.length <= 2 ? `Required Length` : true

The first check checks to see if there is any data in the input and reports 'Required'.

The second check handles the 'clearable' exception and reports no results.

The third check performs a length check for example and reports 'Required Length'.

---
The important thing is the first and second checks which MUST be performed in the correct order.

--------------------------------------------------
modified:   src/views/   TodoView.vue
--------------------------------------------------