-------------------------------------------------- Version:   v3a-todo-theme-v2 Date:      Friday, November 17, 2023 @ 3:08 pm --------------------------------------------------
NOTE: This commit documents a series of previous commits made to TodoView.vue.

Modify the styling of the todo data so that it is visible in Light and Dark Themes.
This was an issue on MOB devices, in the LIVE site.
The issue did not appear during development even thought it was checked.

--------------------------------------------------
modified:   src/views/.  TodoView.vue
--------------------------------------------------

Replace style="color: initial;" with style="color:rgb(var(--v-theme-border))".
-- "color: initial;" didnt seem to work.
-- This failure only presented on the LIVE Site on a MOB device.

This change was applied to the TODO LIST CONTAINER and ZOOM_CONTAINER where appropriate.